+++ src/app/features/cart/cart.module.ts
@@ -3,6 +3,7 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { CartRoutingModule } from './cart-routing.module';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 const LOCAL_COMPONENTS = [
   CartItemComponent
@@ -16,6 +17,8 @@ const LOCAL_COMPONENTS = [
   imports: [
     SharedModule,
     FontAwesomeModule,
+    FormsModule, 
+    ReactiveFormsModule,
     CartRoutingModule,
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]