+++ src/app/shp.module.ts
@@ -14,7 +14,6 @@ import { OrdersModule } from './features/orders/orders.module';
 import { CoreModule } from './core/core.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AdminModule } from './features/admin/admin.module';
-import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { RootStoreModule } from './core/@ngrx/root-store.module';