+++ changelog.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.8.0] - 2021-11-04
+### Removed
+### Changed
+- CartService: isEmptyCart now Subject
+- OrderActivateGuard to work with Observable/Subject
+- ProcessOrder Component
+### Added
+- getKeyValue function
+- EmailDirective validator
+- Add / Remove phone feature
+
 ## [0.7.0] - 2021-10-31
 ### Removed
 ### Changed